A force reinstallation of the whole bootstrap after removing files first.
use this once to install cydia and then delete it. As you can now use the electra app to re enable your jailbreak again. this is created to reset the clock in cydia in case, you delete something by accident and cydia just stops working. Do not use this every time you need to rejailbreak as it will wipe out files and folders and make you redownload all your tweaks on every run, well sort of redownload them so that cydia sees that you installed them and can remove them if needed, Properly.