Code i added for removal of all topanga files before payload is extracted
had to update the branch after figuring out how to creat multiple copies of the same project with minor differences for everyone. This branch if for Initilal jalibreak to install cydia, If you cannot use Electra Jailbreak due to the Topanga error.  added code is in the commits, i have to just add some comments i think on top of what i added